split class and move it into src dir
